Pull out products related functionality into a separate module
