+++ src/modules/products/components/ProductSelect.js
