+++ src/modules/products/components/ProductTable.js
@@ -9,10 +9,10 @@ import IconButton from "@material-ui/core/IconButton";
 import EditIcon from "@material-ui/icons/Edit";
 import DeleteIcon from "@material-ui/icons/Delete";
 
-import EditProductDialog from "containers/components/EditProductDialog";
-import ProductDeleteConfirm from "containers/components/ProductDeleteConfirm.js";
+import Progress from "components/Progress.js";
 
-import Progress from "./Progress.js";
+import EditProductDialog from "../containers/EditProductDialog";
+import ProductDeleteConfirm from "../containers/ProductDeleteConfirm.js";
 
 const styles = theme => ({
   root: {
