+++ src/modules/products/containers/AddProductDialog.js
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 
-import ProductDialog from "components/ProductDialog";
-import { productAdd } from "actions/product";
+import ProductDialog from "../components/ProductDialog";
+import { productAdd } from "../actions";
 
 const mapDispatchToProps = { productAdd };
 const mergeProps = (stateProps, { productAdd }, { onClose, ...props }) => ({
