+++ src/modules/products/containers/ProductTable.js
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 
-import ProductTableWrapper from "components/ProductTable";
-import { productDeleteConfirm, productList } from "actions/product";
+import ProductTableWrapper from "../components/ProductTable";
+import { productDeleteConfirm, productList } from "../actions";
 
 const mapStateToProps = ({ products }) => ({ products });
 const mapDispatchToProps = { productDeleteConfirm, productList };
