+++ src/modules/products/index.js
@@ -1,0 +1,2 @@
+export { default as ProductSelect } from "./components/ProductSelect";
+export { default as ProductsPage } from "./pages/ProductsPage";
