+++ src/reducers/index.js
@@ -1,10 +1,10 @@
 import { combineReducers } from "redux";
 
 import user from "modules/auth/reducer";
+import products from "modules/products/reducer";
 
 import profile from "./profile";
 import boxes from "./boxes";
-import products from "./products";
 
 export default combineReducers({
   products,
